Updates and changes header component
