+++ src/components/Cart.jsx
@@ -1,13 +1,16 @@
 import { Link } from "react-router-dom";
 import useCounterStore from "../hooks/Store";
+import { ShoppingBag } from "lucide-react";
 
 function Cart() {
   const totalCount = useCounterStore((state) => state.totalCount(state));
   
   return (
-    <div className="order-4 md:order-5 relative">
-      <Link to="/checkout" title="Shopping cart" className=""><img src="/assets/shopping-bag.svg" alt="shopping cart icon" className="h-7" /></Link>
-      <div className="flex justify-center absolute left-5 bottom-2 p-2 size-5 rounded-full bg-red-500 text-xs items-center">{totalCount}</div>
+    <div className="order-4 md:order-5 relative mr-4">
+      <Link to="/checkout" title="Shopping cart" className="">
+        <ShoppingBag color="white" size="25" strokeWidth={1}/>
+        <div className="flex justify-center absolute left-4 -bottom-2 p-2 size-5 rounded-full bg-red-600 text-xs items-center">{totalCount}</div>
+      </Link>
     </div>
   )
 }
