+++ src/components/icons.jsx
@@ -19,3 +19,4 @@ export function Email({size = 24, classes}) {
   </svg>
 };
 
+
